backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elevenlabs>=2.27.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.2.1",
]
